requirements.txt
ebooklib
beautifulsoup4
python-dotenv
google-generativeai